Implement AuthView to encapsulate SignIn and SignUp functionality
